Move user and location repeaters to generic places
