+++ corehq/apps/repeaters/models.py
@@ -8,7 +8,10 @@ from django.utils.translation import ugettext_lazy as _
 from requests.auth import HTTPBasicAuth, HTTPDigestAuth
 
 from requests.exceptions import Timeout, ConnectionError
+from corehq.toggles import BETS_INTEGRATION
 from corehq.apps.cachehq.mixins import QuickCachedDocumentMixin
+from corehq.apps.locations.models import SQLLocation
+from corehq.apps.users.models import CommCareUser
 from corehq.form_processor.exceptions import XFormNotFound
 from corehq.util.datadog.metrics import REPEATER_ERROR_COUNT
 from corehq.util.datadog.gauges import datadog_counter
@@ -465,6 +468,42 @@ class AppStructureRepeater(Repeater):
 
     def payload_doc(self, repeat_record):
         return None
+
+
+class BETSUserRepeater(Repeater):
+    friendly_name = _("Forward Users")
+
+    class Meta(object):
+        app_label = 'repeaters'
+
+    @memoized
+    def payload_doc(self, repeat_record):
+        return CommCareUser.get(repeat_record.payload_id)
+
+    @classmethod
+    def available_for_domain(cls, domain):
+        return BETS_INTEGRATION.enabled(domain)
+
+    def __unicode__(self):
+        return "forwarding users to: %s" % self.url
+
+
+class BETSLocationRepeater(Repeater):
+    friendly_name = _("Forward Locations")
+
+    class Meta(object):
+        app_label = 'repeaters'
+
+    @memoized
+    def payload_doc(self, repeat_record):
+        return SQLLocation.objects.get(location_id=repeat_record.payload_id)
+
+    @classmethod
+    def available_for_domain(cls, domain):
+        return BETS_INTEGRATION.enabled(domain)
+
+    def __unicode__(self):
+        return "forwarding locations to: %s" % self.url
 
 
 class RepeatRecord(Document):
