+++ corehq/apps/repeaters/repeater_generators.py
@@ -6,8 +6,16 @@ from uuid import uuid4
 from django.core.serializers.json import DjangoJSONEncoder
 from casexml.apps.case.xform import cases_referenced_by_xform
 
-from corehq.apps.repeaters.models import FormRepeater, CaseRepeater, ShortFormRepeater, \
-    AppStructureRepeater, GeneratorCollection
+from corehq.apps.api.resources.v0_5 import CommCareUserResource
+from corehq.apps.repeaters.models import (
+    FormRepeater,
+    CaseRepeater,
+    ShortFormRepeater,
+    AppStructureRepeater,
+    GeneratorCollection,
+    BETSUserRepeater,
+    BETSLocationRepeater,
+)
 
 from casexml.apps.case.xml import V2
 
@@ -212,3 +220,27 @@ class FormRepeaterJsonPayloadGenerator(BasePayloadGenerator):
 
     def get_test_payload(self, domain):
         return self.get_payload(None, _get_test_form(domain))
+
+
+@RegisterGenerator(BETSUserRepeater, "user_json", "JSON", is_default=True)
+class BETSUserPayloadGenerator(BasePayloadGenerator):
+
+    @property
+    def content_type(self):
+        return 'application/json'
+
+    def get_payload(self, repeat_record, user):
+        resource = CommCareUserResource(api_name='v0.5')
+        bundle = resource.build_bundle(obj=user)
+        return resource.full_dehydrate(bundle).data
+
+
+@RegisterGenerator(BETSLocationRepeater, "user_json", "JSON", is_default=True)
+class BETSLocationPayloadGenerator(BasePayloadGenerator):
+
+    @property
+    def content_type(self):
+        return 'application/json'
+
+    def get_payload(self, repeat_record, location):
+        return location.to_json()
