+++ corehq/apps/repeaters/signals.py
@@ -1,7 +1,13 @@
+from django.db.models.signals import post_save
+from django.dispatch import receiver
+
 from casexml.apps.case.models import CommCareCase
-from corehq.form_processor.models import CommCareCaseSQL
 from casexml.apps.case.signals import case_post_save
 from couchforms.signals import successful_form_received
+
+from corehq.apps.locations.models import SQLLocation
+from corehq.apps.users.signals import commcare_user_post_save
+from corehq.form_processor.models import CommCareCaseSQL
 
 
 def create_form_repeat_records(sender, xform, **kwargs):
@@ -25,6 +31,20 @@ def create_repeat_records(repeater_cls, payload):
         repeaters = repeater_cls.by_domain(domain)
         for repeater in repeaters:
             repeater.register(payload)
+
+
+@receiver(commcare_user_post_save, dispatch_uid="create_bets_user_repeat_records")
+def create_user_repeat_records(sender, couch_user, **kwargs):
+    from corehq.apps.repeaters.models import BETSUserRepeater
+    create_repeat_records(BETSUserRepeater, couch_user)
+
+
+@receiver(post_save, sender=SQLLocation, dispatch_uid="create_bets_location_repeat_records")
+def create_location_repeat_records(sender, raw=False, **kwargs):
+    from corehq.apps.repeaters.models import BETSLocationRepeater
+    if raw:
+        return
+    create_repeat_records(BETSLocationRepeater, kwargs['instance'])
 
 
 successful_form_received.connect(create_form_repeat_records)
