+++ corehq/apps/repeaters/views.py
@@ -36,6 +36,26 @@ class AddCaseRepeaterView(AddRepeaterView):
         return repeater
 
 
+class BETSUserRepeaterView(AddRepeaterView):
+    urlname = "user_repeater"
+    page_title = "Users"
+    page_name = page_title
+
+    @property
+    def page_url(self):
+        return reverse(self.urlname, args=[self.domain])
+
+
+class BETSLocationRepeaterView(AddRepeaterView):
+    urlname = "location_repeater"
+    page_title = "Locations"
+    page_name = page_title
+
+    @property
+    def page_url(self):
+        return reverse(self.urlname, args=[self.domain])
+
+
 class RepeatRecordView(LoginAndDomainMixin, View):
 
     urlname = 'repeat_record'
